Added data types

+++ julia/diffusion.jl
@@ -10,21 +10,21 @@
 	Accessed Friday, November 8th, 2019
 =#
 
-maxsize = 10
-diffusion_coefficient = 0.175
-room_dimension = 5
-speed_of_gas_molecules = 250.0
-timestep = (room_dimension / speed_of_gas_molecules) / maxsize
-distance_between_blocks = room_dimension / maxsize
-dterm = diffusion_coefficient * timestep / (distance_between_blocks * distance_between_blocks)
+maxsize = 10::Int64
+diffusion_coefficient = 0.175::Float64
+room_dimension = 5::Float64
+speed_of_gas_molecules = 250.0::Float64
+timestep = ((room_dimension / speed_of_gas_molecules) / maxsize)::Float64
+distance_between_blocks = (room_dimension / maxsize)::Float64
+dterm = (diffusion_coefficient * timestep / (distance_between_blocks * distance_between_blocks))::Float64
 
-pass = 0
-timeamount = 0.0
-ratioamount = 0.0
+pass = 0::Int64
+timeamount = 0.0::Float64
+ratioamount = 0.0::Float64
 
-timeamount = timeamount + timestep
+timeamount = (timeamount + timestep)::Float64
 
-sumval = 0.0
+sumval = (0.0)::Float64
 
 println(timeamount, " ", ratioamount, " ", sumval)
